Pass missing config file parameter to internal Init() call
This fix also updates skv_base_test's Init() call in order to respect
the command line option to change the path to the configuration file.
Note that the ifndef SKV_CLIENT_UNI has been removed from the Init()
call since: 1) The ifndef should wrap the whole call to Init() since the
number of parameters can change depending on the version, and there are
optional parameters as well, which can lead to conflicts; and 2)
skv_base_test is configured to be compiled in non-MPI version only
anyway.